Update dbaas.md
This needs a bit more cleanup...for instance the instructions for minikube (now) and EKS have the link to add the k8s cluster to PMM however GKS doesn't have a link but instead shows it as part of the screenshots...so steps 13, 14 and 15 don't follow the same logical flow. 

I also see we've duplicated (maybe even triplicated" the steps to run pmm-server in a docker container...this needs MASSIVE CLEANUP PLEASE...all I've done is fix it to make it actually work.