Adding an example alert template to documentation
There is no easy way to create an alert template as we provide no guide so until you can view/clone existing templates we need at least something on the website to help users create a template.  This doesn't even scratch the surface though of how to actually create an alert on something unique nor a guide to what the possible variables/substitutions are.