modify the sentence "Silences specify periods of time to suppress & notifications. " [add alerts]
original: ### Silences
Silences specify periods of time to suppress notifications.... During a silence, PMM continues to track metrics and trigger alerts but does not send notifications to the specified contact points. Once the specified silence expires, notifications are resumed.

Changed to: Silences specify periods of time to suppress alerts and their associated notifications...Once the silence expires alerts and notifications will resume.
It is not precise to say they only silence notifications (by omitting the ref to alerts)....really, they silence alerts, which will consequently send notifications (and sometimes not). This is not the same as saying silences only affect notifications, bc the fired alerts seen in fired alert page and alert rules page are also silenced, which we havent mentioned and is a bit misleading.